requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
Pillow==10.0.0
stripe==5.4.0
django-filter==25.1
drf-spectacular==0.28.0
fcm-django==2.2.1
django-ratelimit==4.1.0 